Issue #116: Develop OS Command Extension
* Developed Os Command Extension
* Made all the necessary changes on metricshub engine and agent
* Tested the code using the metricshub CLI and the agent
