--- metricshub-agent/src/main/java/org/sentrysoftware/metricshub/agent/config/protocols/OsCommandProtocolConfig.java
@@ -1,80 +1,0 @@
-package org.sentrysoftware.metricshub.agent.config.protocols;
-
-/*-
- * ╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲
- * MetricsHub Agent
- * ჻჻჻჻჻჻
- * Copyright 2023 - 2024 Sentry Software
- * ჻჻჻჻჻჻
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- * ╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱
- */
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import java.util.HashSet;
-import java.util.Set;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Builder.Default;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import org.sentrysoftware.metricshub.engine.configuration.IConfiguration;
-import org.sentrysoftware.metricshub.engine.configuration.OsCommandConfiguration;
-import org.sentrysoftware.metricshub.engine.deserialization.TimeDeserializer;
-
-/**
- * Configuration class for the OS command protocol.
- * Extends {@link AbstractProtocolConfig}.
- */
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@EqualsAndHashCode(callSuper = false)
-public class OsCommandProtocolConfig extends AbstractProtocolConfig {
-
-	private boolean useSudo;
-
-	@Default
-	private Set<String> useSudoCommands = new HashSet<>();
-
-	@Default
-	private String sudoCommand = "sudo";
-
-	@Default
-	@JsonDeserialize(using = TimeDeserializer.class)
-	private Long timeout = 120L;
-
-	/**
-	 * Create a new {@link OsCommandConfiguration} instance based on the current members
-	 *
-	 * @return The {@link OsCommandConfiguration} instance
-	 */
-	@Override
-	public IConfiguration toConfiguration() {
-		return OsCommandConfiguration
-			.builder()
-			.useSudo(useSudo)
-			.useSudoCommands(useSudoCommands)
-			.sudoCommand(sudoCommand)
-			.timeout(timeout)
-			.build();
-	}
-
-	@Override
-	public String toString() {
-		return "Local Commands";
-	}
-}
