+++ metricshub-agent/src/main/java/org/sentrysoftware/metricshub/agent/config/protocols/ProtocolsConfig.java
@@ -50,7 +50,8 @@ public class ProtocolsConfig {
 	private IpmiProtocolConfig ipmi;
 
 	@JsonSetter(nulls = SKIP)
-	private SshProtocolConfig ssh;
+	@JsonDeserialize(using = ExtensionConfigDeserializer.class)
+	private IConfiguration ssh;
 
 	@JsonSetter(nulls = SKIP)
 	private WbemProtocolConfig wbem;
@@ -62,7 +63,8 @@ public class ProtocolsConfig {
 	private HttpProtocolConfig http;
 
 	@JsonSetter(nulls = SKIP)
-	private OsCommandProtocolConfig osCommand;
+	@JsonDeserialize(using = ExtensionConfigDeserializer.class)
+	private IConfiguration osCommand;
 
 	@JsonSetter(nulls = SKIP)
 	private WinRmProtocolConfig winrm;
