--- metricshub-agent/src/main/java/org/sentrysoftware/metricshub/agent/config/protocols/SshProtocolConfig.java
@@ -1,91 +1,0 @@
-package org.sentrysoftware.metricshub.agent.config.protocols;
-
-/*-
- * ╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲
- * MetricsHub Agent
- * ჻჻჻჻჻჻
- * Copyright 2023 - 2024 Sentry Software
- * ჻჻჻჻჻჻
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- * ╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱
- */
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import java.io.File;
-import java.util.HashSet;
-import java.util.Set;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Builder.Default;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import org.sentrysoftware.metricshub.engine.configuration.IConfiguration;
-import org.sentrysoftware.metricshub.engine.configuration.SshConfiguration;
-import org.sentrysoftware.metricshub.engine.deserialization.TimeDeserializer;
-
-/**
- * Configuration class for the SSH protocol.
- * Extends {@link AbstractProtocolConfig}.
- */
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@EqualsAndHashCode(callSuper = false)
-public class SshProtocolConfig extends AbstractProtocolConfig {
-
-	@Default
-	@JsonDeserialize(using = TimeDeserializer.class)
-	private Long timeout = 120L;
-
-	private String username;
-	private char[] password;
-	private File privateKey;
-	private boolean useSudo;
-
-	@Default
-	private Set<String> useSudoCommands = new HashSet<>();
-
-	@Default
-	private String sudoCommand = "sudo";
-
-	/**
-	 * Create a new {@link SshConfiguration} instance based on the current members
-	 *
-	 * @return The {@link SshConfiguration} instance
-	 */
-	@Override
-	public IConfiguration toConfiguration() {
-		return SshConfiguration
-			.sshConfigurationBuilder()
-			.username(username)
-			.password(super.decrypt(password))
-			.privateKey(privateKey)
-			.timeout(timeout)
-			.useSudo(useSudo)
-			.useSudoCommands(useSudoCommands)
-			.sudoCommand(sudoCommand)
-			.build();
-	}
-
-	@Override
-	public String toString() {
-		String desc = "SSH";
-		if (username != null) {
-			desc = desc + " as " + username;
-		}
-		return desc;
-	}
-}
