+++ metricshub-agent/src/test/java/org/sentrysoftware/metricshub/agent/helper/ConfigHelperTest.java
@@ -296,21 +296,6 @@ class ConfigHelperTest {
 	}
 
 	@Test
-	void testValidateSshInfo() {
-		assertThrows(InvalidConfigurationException.class, () -> ConfigHelper.validateSshInfo(RESOURCE_KEY, "", 60L));
-		assertThrows(InvalidConfigurationException.class, () -> ConfigHelper.validateSshInfo(RESOURCE_KEY, null, 60L));
-		assertThrows(
-			InvalidConfigurationException.class,
-			() -> ConfigHelper.validateSshInfo(RESOURCE_KEY, USERNAME_CONFIG_VALUE, -60L)
-		);
-		assertThrows(
-			InvalidConfigurationException.class,
-			() -> ConfigHelper.validateSshInfo(RESOURCE_KEY, USERNAME_CONFIG_VALUE, null)
-		);
-		assertDoesNotThrow(() -> ConfigHelper.validateSshInfo(RESOURCE_KEY, USERNAME_CONFIG_VALUE, 60L));
-	}
-
-	@Test
 	void testValidateWbemInfo() {
 		assertThrows(
 			InvalidConfigurationException.class,
@@ -365,13 +350,6 @@ class ConfigHelperTest {
 		assertThrows(InvalidConfigurationException.class, () -> ConfigHelper.validateHttpInfo(RESOURCE_KEY, 60L, null));
 		assertThrows(InvalidConfigurationException.class, () -> ConfigHelper.validateHttpInfo(RESOURCE_KEY, 60L, 66666));
 		assertDoesNotThrow(() -> ConfigHelper.validateHttpInfo(RESOURCE_KEY, 60L, 1234));
-	}
-
-	@Test
-	void testValidateOsCommandInfo() {
-		assertThrows(InvalidConfigurationException.class, () -> ConfigHelper.validateOsCommandInfo(RESOURCE_KEY, -60L));
-		assertThrows(InvalidConfigurationException.class, () -> ConfigHelper.validateOsCommandInfo(RESOURCE_KEY, null));
-		assertDoesNotThrow(() -> ConfigHelper.validateOsCommandInfo(RESOURCE_KEY, 60L));
 	}
 
 	@Test
