+++ metricshub-engine/pom.xml
@@ -98,10 +98,6 @@
 		</dependency>
 		<dependency>
 			<groupId>org.sentrysoftware</groupId>
-			<artifactId>ssh</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.sentrysoftware</groupId>
 			<artifactId>jflat</artifactId>
 		</dependency>
 		<dependency>
