+++ metricshub-engine/src/test/java/org/sentrysoftware/metricshub/engine/configuration/SshTestConfiguration.java
@@ -1,0 +1,95 @@
+package org.sentrysoftware.metricshub.engine.configuration;
+
+/*-
+ * ╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲
+ * MetricsHub OsCommand Extension
+ * ჻჻჻჻჻჻
+ * Copyright 2023 - 2024 Sentry Software
+ * ჻჻჻჻჻჻
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * ╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱
+ */
+
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import java.util.HashSet;
+import java.util.Set;
+import lombok.AllArgsConstructor;
+import lombok.Builder.Default;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import org.sentrysoftware.metricshub.engine.common.exception.InvalidConfigurationException;
+import org.sentrysoftware.metricshub.engine.common.helpers.StringHelper;
+import org.sentrysoftware.metricshub.engine.deserialization.TimeDeserializer;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@SuperBuilder
+@EqualsAndHashCode(callSuper = false)
+public class SshTestConfiguration extends OsCommandTestConfiguration {
+
+	@Default
+	@JsonDeserialize(using = TimeDeserializer.class)
+	private Long timeout = 120L;
+
+	private String username;
+	private char[] password;
+	private String privateKey;
+	private boolean useSudo;
+
+	@Default
+	private Set<String> useSudoCommands = new HashSet<>();
+
+	@Default
+	private String sudoCommand = "sudo";
+
+	@Override
+	public void validateConfiguration(String resourceKey) throws InvalidConfigurationException {
+		StringHelper.validateConfigurationAttribute(
+			username,
+			attr -> attr == null || attr.isBlank(),
+			() ->
+				String.format(
+					"Resource %s - No username configured for protocol %s." +
+					" This resource will not be monitored. Please verify the configured username.",
+					resourceKey,
+					"SSH"
+				)
+		);
+
+		StringHelper.validateConfigurationAttribute(
+			timeout,
+			attr -> attr == null || attr < 0L,
+			() ->
+				String.format(
+					"Resource %s - Timeout value is invalid for protocol %s." +
+					" Timeout value returned: %s. This resource will not be monitored. Please verify the configured timeout value.",
+					resourceKey,
+					"SSH",
+					timeout
+				)
+		);
+	}
+
+	@Override
+	public String toString() {
+		String desc = "SSH";
+		if (username != null) {
+			desc = desc + " as " + username;
+		}
+		return desc;
+	}
+}
