+++ metricshub-engine/src/test/java/org/sentrysoftware/metricshub/engine/strategy/collect/ProtocolHealthCheckStrategyTest.java
@@ -2,7 +2,6 @@ package org.sentrysoftware.metricshub.engine.strategy.collect;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyLong;
@@ -17,7 +16,6 @@ import static org.sentrysoftware.metricshub.engine.constants.Constants.HOSTNAME;
 import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.DOWN;
 import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.HTTP_UP_METRIC;
 import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.IPMI_UP_METRIC;
-import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.SSH_UP_METRIC;
 import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.UP;
 import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.WBEM_TEST_QUERY;
 import static org.sentrysoftware.metricshub.engine.strategy.collect.ProtocolHealthCheckStrategy.WBEM_UP_METRIC;
@@ -46,14 +44,12 @@ import org.sentrysoftware.metricshub.engine.common.exception.ClientException;
 import org.sentrysoftware.metricshub.engine.configuration.HostConfiguration;
 import org.sentrysoftware.metricshub.engine.configuration.HttpConfiguration;
 import org.sentrysoftware.metricshub.engine.configuration.IpmiConfiguration;
-import org.sentrysoftware.metricshub.engine.configuration.SshConfiguration;
 import org.sentrysoftware.metricshub.engine.configuration.TestConfiguration;
 import org.sentrysoftware.metricshub.engine.configuration.WbemConfiguration;
 import org.sentrysoftware.metricshub.engine.configuration.WinRmConfiguration;
 import org.sentrysoftware.metricshub.engine.configuration.WmiConfiguration;
 import org.sentrysoftware.metricshub.engine.extension.ExtensionManager;
 import org.sentrysoftware.metricshub.engine.extension.IProtocolExtension;
-import org.sentrysoftware.metricshub.engine.strategy.utils.OsCommandHelper;
 import org.sentrysoftware.metricshub.engine.telemetry.Monitor;
 import org.sentrysoftware.metricshub.engine.telemetry.TelemetryManager;
 import org.sentrysoftware.wbem.javax.wbem.WBEMException;
@@ -124,37 +120,6 @@ class ProtocolHealthCheckStrategyTest {
 					.hostId(HOSTNAME)
 					.hostname(HOSTNAME)
 					.configurations(Map.of(TestConfiguration.class, TestConfiguration.builder().build()))
-					.build()
-			)
-			.build();
-	}
-
-	/**
-	 * Creates and returns a TelemetryManager instance with an SSH configuration.
-	 *
-	 * @return A TelemetryManager instance configured with an SSH configuration.
-	 */
-	private TelemetryManager createTelemetryManagerWithSshConfig() {
-		// Create a telemetry manager
-		return TelemetryManager
-			.builder()
-			.monitors(monitors)
-			.hostConfiguration(
-				HostConfiguration
-					.builder()
-					.hostId(HOSTNAME)
-					.hostname(HOSTNAME)
-					.configurations(
-						Map.of(
-							SshConfiguration.class,
-							SshConfiguration
-								.sshConfigurationBuilder()
-								.username("username")
-								.password("password".toCharArray())
-								.timeout(60L)
-								.build()
-						)
-					)
 					.build()
 			)
 			.build();
@@ -266,20 +231,6 @@ class ProtocolHealthCheckStrategyTest {
 			.build();
 	}
 
-	/**
-	 * Creates and returns a TelemetryManager instance without any configuration.
-	 *
-	 * @return A TelemetryManager instance.
-	 */
-	private TelemetryManager createTelemetryManagerWithoutConfig() {
-		// Create a telemetry manager
-		return TelemetryManager
-			.builder()
-			.monitors(monitors)
-			.hostConfiguration(HostConfiguration.builder().hostId(HOSTNAME).hostname(HOSTNAME).build())
-			.build();
-	}
-
 	@Test
 	void testCheckHttpDownHealth() {
 		// Create a telemetry manager using an HTTP HostConfiguration.
@@ -349,224 +300,6 @@ class ProtocolHealthCheckStrategyTest {
 			extensionManager
 		);
 		assertDoesNotThrow(() -> healthCheckStrategy.run());
-	}
-
-	@Test
-	void testCheckSshHealthLocally() {
-		// Create a telemetry manager using an SSH HostConfiguration.
-		final TelemetryManager telemetryManager = createTelemetryManagerWithSshConfig();
-
-		// Create a new protocol health check strategy
-		final ProtocolHealthCheckStrategy sshHealthCheckStrategy = new ProtocolHealthCheckStrategy(
-			telemetryManager,
-			CURRENT_TIME_MILLIS,
-			clientsExecutorMock,
-			ExtensionManager.empty()
-		);
-
-		// Setting SSH host properties
-		telemetryManager.getHostProperties().setMustCheckSshStatus(true);
-		telemetryManager.getHostProperties().setOsCommandExecutesLocally(true);
-
-		try (MockedStatic<OsCommandHelper> staticOsCommandHelper = Mockito.mockStatic(OsCommandHelper.class)) {
-			staticOsCommandHelper
-				.when(() -> OsCommandHelper.runLocalCommand(anyString(), anyLong(), any()))
-				.thenReturn(SUCCESS_RESPONSE);
-
-			// Start the SSH Health Check strategy
-			sshHealthCheckStrategy.run();
-
-			assertEquals(UP, telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC).getValue());
-		}
-
-		try (MockedStatic<OsCommandHelper> staticOsCommandHelper = Mockito.mockStatic(OsCommandHelper.class)) {
-			staticOsCommandHelper.when(() -> OsCommandHelper.runLocalCommand(anyString(), anyLong(), any())).thenReturn(null);
-
-			// Start the SSH Health Check strategy
-			sshHealthCheckStrategy.run();
-
-			assertEquals(DOWN, telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC).getValue());
-		}
-	}
-
-	@Test
-	void testCheckSshUpHealthRemotely() {
-		// Create a telemetry manager using an SSH HostConfiguration.
-		final TelemetryManager telemetryManager = createTelemetryManagerWithSshConfig();
-
-		// Create a new protocol health check strategy
-		final ProtocolHealthCheckStrategy sshHealthCheckStrategy = new ProtocolHealthCheckStrategy(
-			telemetryManager,
-			CURRENT_TIME_MILLIS,
-			clientsExecutorMock,
-			ExtensionManager.empty()
-		);
-
-		// Setting SSH host properties
-		telemetryManager.getHostProperties().setMustCheckSshStatus(true);
-		telemetryManager.getHostProperties().setOsCommandExecutesLocally(false);
-		telemetryManager.getHostProperties().setOsCommandExecutesRemotely(true);
-
-		try (MockedStatic<OsCommandHelper> staticOsCommandHelper = Mockito.mockStatic(OsCommandHelper.class)) {
-			staticOsCommandHelper
-				.when(() ->
-					OsCommandHelper.runSshCommand(anyString(), anyString(), any(SshConfiguration.class), anyLong(), any(), any())
-				)
-				.thenReturn(SUCCESS_RESPONSE);
-
-			// Start the SSH Health Check strategy
-			sshHealthCheckStrategy.run();
-
-			assertEquals(UP, telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC).getValue());
-		}
-
-		try (MockedStatic<OsCommandHelper> staticOsCommandHelper = Mockito.mockStatic(OsCommandHelper.class)) {
-			staticOsCommandHelper
-				.when(() ->
-					OsCommandHelper.runSshCommand(anyString(), anyString(), any(SshConfiguration.class), anyLong(), any(), any())
-				)
-				.thenReturn(null);
-
-			// Start the SSH Health Check strategy
-			sshHealthCheckStrategy.run();
-
-			assertEquals(DOWN, telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC).getValue());
-		}
-	}
-
-	@Test
-	void testCheckSshUpHealthBothLocallyAndRemotely() {
-		// Create a telemetry manager using an SSH HostConfiguration.
-		final TelemetryManager telemetryManager = createTelemetryManagerWithSshConfig();
-
-		// Create a new protocol health check strategy
-		final ProtocolHealthCheckStrategy sshHealthCheckStrategy = new ProtocolHealthCheckStrategy(
-			telemetryManager,
-			CURRENT_TIME_MILLIS,
-			clientsExecutorMock,
-			ExtensionManager.empty()
-		);
-
-		// Setting SSH host properties
-		telemetryManager.getHostProperties().setMustCheckSshStatus(true);
-		telemetryManager.getHostProperties().setOsCommandExecutesLocally(true);
-		telemetryManager.getHostProperties().setOsCommandExecutesRemotely(true);
-
-		// Both local and remote commands working fine
-		try (MockedStatic<OsCommandHelper> staticOsCommandHelper = Mockito.mockStatic(OsCommandHelper.class)) {
-			staticOsCommandHelper
-				.when(() ->
-					OsCommandHelper.runSshCommand(anyString(), anyString(), any(SshConfiguration.class), anyLong(), any(), any())
-				)
-				.thenReturn(SUCCESS_RESPONSE);
-
-			staticOsCommandHelper
-				.when(() -> OsCommandHelper.runLocalCommand(anyString(), anyLong(), any()))
-				.thenReturn(SUCCESS_RESPONSE);
-
-			// Start the SSH Health Check strategy
-			sshHealthCheckStrategy.run();
-
-			assertEquals(UP, telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC).getValue());
-		}
-
-		// Local commands not working
-		try (MockedStatic<OsCommandHelper> staticOsCommandHelper = Mockito.mockStatic(OsCommandHelper.class)) {
-			staticOsCommandHelper
-				.when(() ->
-					OsCommandHelper.runSshCommand(anyString(), anyString(), any(SshConfiguration.class), anyLong(), any(), any())
-				)
-				.thenReturn(SUCCESS_RESPONSE);
-
-			staticOsCommandHelper.when(() -> OsCommandHelper.runLocalCommand(anyString(), anyLong(), any())).thenReturn(null);
-
-			// Start the SSH Health Check strategy
-			sshHealthCheckStrategy.run();
-
-			assertEquals(DOWN, telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC).getValue());
-		}
-		// remote command not working
-		try (MockedStatic<OsCommandHelper> staticOsCommandHelper = Mockito.mockStatic(OsCommandHelper.class)) {
-			staticOsCommandHelper
-				.when(() ->
-					OsCommandHelper.runSshCommand(anyString(), anyString(), any(SshConfiguration.class), anyLong(), any(), any())
-				)
-				.thenReturn(null);
-
-			staticOsCommandHelper
-				.when(() -> OsCommandHelper.runLocalCommand(anyString(), anyLong(), any()))
-				.thenReturn(SUCCESS_RESPONSE);
-
-			// Start the SSH Health Check strategy
-			sshHealthCheckStrategy.run();
-
-			assertEquals(DOWN, telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC).getValue());
-		}
-
-		// Both local and remote commands not working, but not throwing exceptions
-		try (MockedStatic<OsCommandHelper> staticOsCommandHelper = Mockito.mockStatic(OsCommandHelper.class)) {
-			staticOsCommandHelper
-				.when(() ->
-					OsCommandHelper.runSshCommand(anyString(), anyString(), any(SshConfiguration.class), anyLong(), any(), any())
-				)
-				.thenReturn(null);
-
-			staticOsCommandHelper.when(() -> OsCommandHelper.runLocalCommand(anyString(), anyLong(), any())).thenReturn(null);
-
-			// Start the SSH Health Check strategy
-			sshHealthCheckStrategy.run();
-
-			assertEquals(DOWN, telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC).getValue());
-		}
-	}
-
-	@Test
-	void testCheckSshNoHealthWhenMustCheckFalse() {
-		// Create a telemetry manager using an SSH HostConfiguration.
-		final TelemetryManager telemetryManager = createTelemetryManagerWithSshConfig();
-
-		// Create a new protocol health check strategy
-		final ProtocolHealthCheckStrategy sshHealthCheckStrategy = new ProtocolHealthCheckStrategy(
-			telemetryManager,
-			CURRENT_TIME_MILLIS,
-			clientsExecutorMock,
-			ExtensionManager.empty()
-		);
-
-		// Setting SSH host properties
-		telemetryManager.getHostProperties().setMustCheckSshStatus(false);
-		telemetryManager.getHostProperties().setOsCommandExecutesLocally(true);
-		telemetryManager.getHostProperties().setOsCommandExecutesRemotely(true);
-
-		// Start the SSH Health Check strategy
-		sshHealthCheckStrategy.run();
-
-		assertNull(telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC));
-	}
-
-	@Test
-	void testCheckSshNoHealthWhenNoConfiguration() {
-		// Create a telemetry manager without configuration.
-		final TelemetryManager telemetryManager = createTelemetryManagerWithoutConfig();
-
-		// Create a new protocol health check strategy
-		final ProtocolHealthCheckStrategy sshHealthCheckStrategy = new ProtocolHealthCheckStrategy(
-			telemetryManager,
-			CURRENT_TIME_MILLIS,
-			clientsExecutorMock,
-			ExtensionManager.empty()
-		);
-
-		// Setting SSH host properties
-		telemetryManager.getHostProperties().setMustCheckSshStatus(false);
-		telemetryManager.getHostProperties().setOsCommandExecutesLocally(true);
-		telemetryManager.getHostProperties().setOsCommandExecutesRemotely(true);
-
-		// Start the SSH Health Check strategy
-		sshHealthCheckStrategy.run();
-
-		// make sure that SSH health check is not performed if an SSH config is not present
-		assertNull(telemetryManager.getEndpointHostMonitor().getMetric(SSH_UP_METRIC));
 	}
 
 	@Test
