+++ metricshub-linux/pom.xml
@@ -25,6 +25,11 @@
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>metricshub-oscommand-extension</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+		<dependency>
 			<groupId>org.sentrysoftware</groupId>
 			<artifactId>metricshub-jre-linux</artifactId>
 			<version>${metricshub-jre.version}</version>
@@ -91,6 +96,11 @@
 								<artifactItem>
 									<groupId>${project.groupId}</groupId>
 									<artifactId>metricshub-snmp-extension</artifactId>
+									<version>${project.version}</version>
+								</artifactItem>
+								<artifactItem>
+									<groupId>${project.groupId}</groupId>
+									<artifactId>metricshub-oscommand-extension</artifactId>
 									<version>${project.version}</version>
 								</artifactItem>
 								<artifactItem>
