+++ metricshub-oscommand-extension/src/main/java/lombok.config
@@ -1,0 +1,2 @@
+lombok.nonNull.exceptionType = IllegalArgumentException
+lombok.addLombokGeneratedAnnotation = true
