+++ metricshub-oscommand-extension/src/main/resources/META-INF/services/org.sentrysoftware.metricshub.engine.extension.IProtocolExtension
@@ -1,0 +1,1 @@
+org.sentrysoftware.metricshub.extension.oscommand.OsCommandExtension
