+++ metricshub-oscommand-extension/src/test/java/org/sentrysoftware/metricshub/extension/oscommand/OsCommandConfigurationTest.java
@@ -1,7 +1,6 @@
-package org.sentrysoftware.metricshub.engine.configuration;
+package org.sentrysoftware.metricshub.extension.oscommand;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.sentrysoftware.metricshub.engine.constants.Constants.OS_COMMAND_CONFIGURATION_TO_STRING;
 
 import org.junit.jupiter.api.Test;
 
@@ -9,6 +8,8 @@ import org.junit.jupiter.api.Test;
  * Test of {@link OsCommandConfiguration}
  */
 class OsCommandConfigurationTest {
+
+	public static final String OS_COMMAND_CONFIGURATION_TO_STRING = "Local Commands";
 
 	@Test
 	void testToString() {
