+++ metricshub-windows/pom.xml
@@ -27,6 +27,11 @@
 		</dependency>
 		<dependency>
 			<groupId>${project.groupId}</groupId>
+			<artifactId>metricshub-oscommand-extension</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
 			<artifactId>metricshub-classloader-agent</artifactId>
 			<version>${project.version}</version>
 		</dependency>
@@ -97,6 +102,11 @@
 								<artifactItem>
 									<groupId>${project.groupId}</groupId>
 									<artifactId>metricshub-snmp-extension</artifactId>
+									<version>${project.version}</version>
+								</artifactItem>
+								<artifactItem>
+									<groupId>${project.groupId}</groupId>
+									<artifactId>metricshub-oscommand-extension</artifactId>
 									<version>${project.version}</version>
 								</artifactItem>
 								<artifactItem>
